requirements.txt

Django[argon2]==3.2.17
wagtail==4.2.2

# External Libraries
requests==2.28.2

# Templates
django-compressor==2.4.1
django-libsass==0.9
wagtailfontawesome==1.2.1

# Scheduled tasks
django-kronos==1.0

# User management
rules==3.3
django-simple-email-confirmation==0.70

# Logging
sentry-sdk==1.15.0

# Postgres
psycopg2==2.9.5

# Phonenumber validation
phonenumbers==8.13.6

# Wagtail media module
wagtailmedia==0.9.0

djangorestframework==3.13.1

# Recaptcha for wagtail form pages
wagtail-django-recaptcha==1.0

django-jsonschema-form==1.0.3

# Improved environment variable handling
python-decouple==3.7
